migrate 3 more tests across
